No longer depend on plugin-profiles
I believe for scala-eclipse template it is not useful to depend on the
plugin-profiles, as we do not need the dependency on scala-ide.

This commit also changes the build to use plain maven artifacts instead
of using the scala-ide p2 repo. I'm not sure this improves anything, but
I'd like to give it a try.

Bundle the scala-library together with the plug-in

Instead of requiring the scala library to be available to install the
plug-in, it is much better to package it together with the plugin.

The rational is that this template is used for creating an Eclipse
plug-in that can be developed using Scala, and there should be no
requirement of having the Scala IDE installed.

Updated README

Removed junit dependency

Also removed the scala-library dependency from the test bundle, as it
is inherited from the plugin MANIFEST.

Removed the appending of the git hash to the plug-in version